updated lecture code to include coordinates
In the Line class for the code the instructor wrote, there was an error message stating that a Vector object is not iterable.  It appears calling the function first_nonzero_ index was the issue.  I added .coordinates to the Vector that was trying to be passed because the coordinates of a Vector object are iterable.